Move score functionality from Course to Module
