+++ src/main/java/seedu/duke/Course.java
@@ -8,14 +8,12 @@ public class Course {
     private final int credits;
     private final String grade ;
     private final boolean isMajor;
-    private final Map<String, Integer> scoreBreakdown;
 
     public Course(String code, int credits, String grade, boolean isMajor) {
         this.code = code;
         this.credits = credits;
         this.grade = grade;
         this.isMajor = isMajor;
-        this.scoreBreakdown = new HashMap<>();
     }
 
     public String getCode() {
@@ -43,18 +41,4 @@ public class Course {
         return code + " | " + credits +  " | " + grade +  " | " + (isMajor ? 1 : 0);
     }
 
-    public void setScoreBreakdown(Map<String, Integer> newBreakdown) {
-        scoreBreakdown.clear();
-        if (newBreakdown != null) {
-            scoreBreakdown.putAll(newBreakdown);
-        }
-    }
-
-    public Map<String, Integer> getScoreBreakdown() {
-        return new HashMap<>(scoreBreakdown);
-    }
-
-    public boolean hasBreakdown() {
-        return !scoreBreakdown.isEmpty();
-    }
 }
